requirements.txt
aiobotocore
paco
retrying
pytest-asyncio
